user token login

+++ src/graphql/user/resolvers.js
@@ -1,5 +1,7 @@
 // QUERIES
 
+import { checkOwner } from '../login/utils/login-functions';
+
 const users = async (_, { input }, { dataSources }) => {
   const users = await dataSources.userApi.getUsers(input);
   return users;
@@ -15,11 +17,17 @@ const createUser = async (_, { data }, { dataSources }) => {
   return dataSources.userApi.createUser(data);
 };
 
-const updateUser = async (_, { userId, data }, { dataSources }) => {
+const updateUser = async (
+  _,
+  { userId, data },
+  { dataSources, loggedUserId },
+) => {
+  checkOwner(userId, loggedUserId);
   return dataSources.userApi.updateUser(userId, data);
 };
 
-const deleteUser = async (_, { userId }, { dataSources }) => {
+const deleteUser = async (_, { userId }, { dataSources, loggedUserId }) => {
+  checkOwner(userId, loggedUserId);
   return dataSources.userApi.deleteUser(userId);
 };
